Backend/requirements.txt
aiomqtt==2.5.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
numpy==2.4.2
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
pymodbus==3.12.0
pyserial==3.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
websockets==16.0
